+++ bcbio/distributed/multitasks.py
@@ -32,6 +32,10 @@ def run_umi_transform(*args):
     return umi.umi_transform(*args)
 
 @utils.map_wrap
+def demultiplex_samples(*args):
+    return umi.demultiplex_samples(*args)
+
+@utils.map_wrap
 def run_kallisto_singlecell(*args):
     return kallisto.run_kallisto_singlecell(*args)
 
@@ -142,10 +146,6 @@ def postprocess_variants(*args):
 @utils.map_wrap
 def pipeline_summary(*args):
     return qcsummary.pipeline_summary(*args)
-
-@utils.map_wrap
-def split_for_qc(*args):
-    return qcsummary.split_for_qc(*args)
 
 @utils.map_wrap
 def qc_to_rec(*args):