+++ bcbio/structural/prioritize.py
@@ -113,7 +113,8 @@ def _prioritize_vcf(caller, vcf_file, prioritize_by, post_prior_fn, work_dir, da
                 do.run(cmd.format(**locals()), "Prioritize: simplified annotation output")
         simple_vcf = vcfutils.bgzip_and_index(vcfutils.sort_by_ref(simple_vcf, data), data["config"])
         with file_transaction(data, out_file) as tx_out_file:
-            cmd = ("zcat {simple_vcf} | vawk -v SNAME={sample} -v CALLER={caller} "
+            export = utils.local_path_export()
+            cmd = ("{export} zcat {simple_vcf} | vawk -v SNAME={sample} -v CALLER={caller} "
                    """'{{if (($7 == "PASS" || $7 == ".") && (S${sample}$GT != "0/0")) """
                    "print CALLER,SNAME,$1,$2,I$END,"
                    """I$SVTYPE=="BND" ? I$SVTYPE":"$3":"I$MATEID : I$SVTYPE,"""